+++ src/redturtle/agidtheme/upgrades/upgrades.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
-from redturtle.agidtheme import logger
 from plone import api
+from redturtle.agidtheme import logger
+
 
 DEFAULT_PROFILE = 'profile-redturtle.agidtheme:default'